requirements.txt
quart
quart-compress
uvloop
hypercorn[uvloop]
google.oauth2
google.auth.transport
time
secrets
asyncio
string
nextcord